+++ start_jupyter_cm/linux.py
@@ -2,15 +2,18 @@ import os, sys
 import stat
 from subprocess import call
 import shutil
+import pathlib
 
 from .utils import get_environment_label
 
-PATH = "%s/bin"%sys.exec_prefix
+
+PATH = f"{sys.exec_prefix}/bin"
 CONDA_ENV_LABEL = get_environment_label()
 
 
-script = \
-"""#!%s
+def get_script(python_exec, path, terminal):
+    script = \
+f"""#!{python_exec}
 
 import sys
 import os.path
@@ -19,18 +22,56 @@ import subprocess
 folders = [path for path in sys.argv[1:] if os.path.isdir(path)]
 any_file_selected = len(folders) < len(sys.argv[1:])
 if any_file_selected:
-    subprocess.Popen(["%s/jupyter-%s"])
+    subprocess.Popen(["{path}/jupyter-{terminal}"])
 for folder in folders:
     os.chdir(folder)
-    subprocess.Popen(["%s/jupyter-%s"])
+    subprocess.Popen(["{path}/jupyter-{terminal}"])
     os.chdir("..")
 
 """
+    return script
+
+def get_desktop(path, terminal, logo, shortcut_name):
+    """
+    Returns the contents of a desktop file that will launch Jupyter
+    from the current environment.
+
+    Parameters
+    ----------
+    path : str
+        'bin' path of the python distribution 
+    terminal : str
+        Either 'notebook', 'qtconsole', or 'lab'
+    logo : str
+        Path to the logo file to use for the .desktop file
+    shortcut_name : str
+        Name of the shortcut.
+    """
+    exec_ = os.path.join(path, f'jupyter-{terminal}')
+    desktop_file = \
+f"""\n
+[Desktop Action {shortcut_name}]\n
+Exec='{exec_}'\n
+Icon={logo}\n
+Name={shortcut_name}\n
+\n
+[Desktop Entry]\n
+Actions={shortcut_name};\n
+MimeType=\n
+ServiceTypes=inode/directory\n
+Type=Service\n
+X-KDE-ServiceTypes=KonqPopupMenu/Plugin,
+inode/directory,all/all,all/allfiles\n
+"""
+    return desktop_file
+
+
+
 
 def check_supported_file_manager(manager_file_path):
     if len(manager_file_path) == 0:
-        print("Nothing done. Currently only 'Nautilus' and 'Caja' are "
-              "supported file manager.")
+        print("Nothing done. Currently only 'Nautilus', 'Caja' and 'Dolphin' "
+              "are supported file manager.")
         return False
     return True
 
@@ -42,10 +83,12 @@ def get_file_manager_config(file_manager=None):
         file_manager_config['nautilus'] = os.path.expanduser("~/.local/share/nautilus")
     if shutil.which("caja"):
         file_manager_config['caja'] = os.path.expanduser("~/.config/caja")
+    if shutil.which("dolphin"):
+        file_manager_config['dolphin'] = os.path.expanduser("~/.local/share/kservices5/ServiceMenus")
 
     if file_manager is not None:
         if file_manager not in file_manager_config.keys():
-            print("File manager '%s' not installed or not supported." % file_manager)
+            print(f"File manager '{file_manager}' not installed or not supported.")
             return {}
         return {file_manager:file_manager_config[file_manager]}
     else:
@@ -65,24 +108,33 @@ def add_jupyter_here(file_manager=None):
 
     python_exec = shutil.which("python")
 
-    for name, path in manager_config_path.items():
-        print("File manager: %s" %name)
-        scripts_folder_path = os.path.join(path, "scripts")
+    for name, scripts_folder_path in manager_config_path.items():
+        print(f"File manager: {name}")
 
-        if not os.path.exists(scripts_folder_path):
-            os.makedirs(scripts_folder_path)
         for terminal in ["qtconsole", "notebook", "lab"]:
-            script_path = os.path.join(scripts_folder_path, "Jupyter %s here%s" % (
-                    terminal, CONDA_ENV_LABEL))
-            if (not os.path.exists(script_path) and
-                shutil.which("jupyter-%s" % terminal)):
+            shortcut_name = f"Jupyter {terminal} here{CONDA_ENV_LABEL}"
+            if name == 'dolphin':
+                script_path = os.path.join(scripts_folder_path, f"{shortcut_name}.desktop")
+                script = get_desktop(PATH, terminal, logos[terminal], shortcut_name)
+            else:
+                script_path = os.path.join(scripts_folder_path, 'scripts', shortcut_name)
+                script = get_script(python_exec, PATH, terminal)
+
+            if not os.path.exists(scripts_folder_path):
+                # In case the parent folder doesn't exist
+                pathlib.Path(scripts_folder_path).mkdir(parents=True)
+
+            # Check that we are getting jupyter from the current environment
+            if PATH in shutil.which(f"jupyter-{terminal}"):
                 with open(script_path, "w") as f:
-                    f.write(script % (python_exec, PATH, terminal, PATH, terminal))
+                    f.write(script)
                 st = os.stat(script_path)
                 os.chmod(script_path, st.st_mode | stat.S_IEXEC)
-                call(['gio', 'set', '-t', 'string', '%s' % script_path,
-                      'metadata::custom-icon', 'file://%s' % logos[terminal]])
-                print('Jupyter %s here%s created.' % (terminal, CONDA_ENV_LABEL))
+                # For nautilus and caja, we need to call gio
+                if name != 'dolphin':
+                    call(['gio', 'set', '-t', 'string', script_path,
+                          'metadata::custom-icon', f'file://{logos[terminal]}'])
+                print(f'{shortcut_name} created.')
 
 
 def remove_jupyter_here(file_manager=None):
@@ -90,12 +142,16 @@ def remove_jupyter_here(file_manager=None):
     if not check_supported_file_manager(manager_config_path):
         return
 
-    for name, path in manager_config_path.items():
-        print("File manager: %s" %name)
-        scripts_folder_path = os.path.join(path, "scripts")
+    for name, scripts_folder_path in manager_config_path.items():
+        print(f"File manager: {name}")
+
         for terminal in ["qtconsole", "notebook", "lab"]:
-            script_path = os.path.join(scripts_folder_path, "Jupyter %s here%s" %(
-                    terminal, CONDA_ENV_LABEL))
+            shortcut_name = f"Jupyter {terminal} here{CONDA_ENV_LABEL}"
+            if name == 'dolphin':
+                script_path = os.path.join(scripts_folder_path, f"{shortcut_name}.desktop")
+            else:
+                script_path = os.path.join(scripts_folder_path, 'scripts', shortcut_name)
+
             if os.path.exists(script_path):
                 os.remove(script_path)
-                print("Jupyter %s here%s removed." % (terminal, CONDA_ENV_LABEL))
+                print(f"{shortcut_name} removed.")